+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Fix bug which would result in crashes if one component had multiple references to the same component instance and more than one references used `@ComponentDependency` annotation.
+
 ### [v0.219](https://github.com/arez/arez/tree/v0.219) (2025-03-31) · [Full Changelog](https://github.com/arez/arez/compare/v0.218...v0.219)
 
 Changes in this release: